Fix empty schema id type mismatch and revert id docstring to UInt64
The id column type depends on the phyloframe version
(UInt32 locally, UInt64 on CI). Remove the UInt32 cast from
_apply_empty_output_schema and skip id type checking in test_empty
and test_single_genome to avoid version-dependent failures.

+++ hstrat/dataframe/_surface_build_tree.py
@@ -115,7 +115,7 @@ def surface_build_tree(
         alife standard format, with the following columns:
 
         Required schema:
-        - 'id' : pl.UInt32
+        - 'id' : pl.UInt64
             - Unique identifier for each taxon (RE alife standard format).
         - 'ancestor_id' : pl.UInt64
             - Unique identifier for ancestor taxon  (RE alife standard format).
